Animations
Using the fadeInOutTo() function to see the box fade in and out and change the opacity.

Using the hideElement() function to see the box go away, reappear, slide up, slide down, and toggle the slide.

Using the chainAnimations() function to link delay, fadeTo and fadeOut together and see the animation.